+++ spec/fixtures/hook/method_named_call.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+class MethodNamedCall
+  def to_s
+    'MethodNamedCall'
+  end
+
+  def call(a, b, c, d, e)
+    [ a, b, c, d, e ].join(' ')
+  end
+end